% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicSymmetricPolynomials.R
\name{cshPolynomial}
\alias{cshPolynomial}
\title{Complete symmetric homogeneous polynomial}
\usage{
cshPolynomial(n, lambda)
}
\arguments{
\item{n}{integer, the number of variables}

\item{lambda}{an integer partition, given as a vector of decreasing
nonnegative integers}
}
\value{
A \code{qspray} object.
}
\description{
Returns a complete symmetric homogeneous polynomial.
}
\examples{
library(jack)
cshPolynomial(3, c(3, 1))
}
