% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESF.R
\name{ESFpoly}
\alias{ESFpoly}
\title{Elementary symmetric function}
\usage{
ESFpoly(m, lambda)
}
\arguments{
\item{m}{integer, the number of variables}

\item{lambda}{an integer partition, given as a vector of decreasing
integers}
}
\value{
A polynomial (\code{mvp} object; see \link[mvp]{mvp-package}).
}
\description{
Returns an elementary symmetric function as a polynomial.
}
\examples{
ESFpoly(3, c(3,1))
}
