% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicSymmetricPolynomials.R
\name{esPolynomial}
\alias{esPolynomial}
\title{Elementary symmetric polynomial}
\usage{
esPolynomial(n, lambda)
}
\arguments{
\item{n}{integer, the number of variables}

\item{lambda}{an integer partition, given as a vector of decreasing
nonnegative integers}
}
\value{
A \code{qspray} object.
}
\description{
Returns an elementary symmetric polynomial.
}
\examples{
library(jack)
esPolynomial(3, c(3, 1))
}
