% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialSchurPol.R
\name{factorialSchurPol}
\alias{factorialSchurPol}
\title{Factorial Schur polynomial}
\usage{
factorialSchurPol(n, lambda, a)
}
\arguments{
\item{n}{number of variables}

\item{lambda}{integer partition}

\item{a}{vector of \code{bigq} numbers, or vector of elements coercible
to \code{bigq} numbers}
}
\value{
A \code{qspray} polynomial.
}
\description{
Computes a factorial Schur polynomial.
}
\examples{
#
}
