% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GreenPolynomials.R
\name{GreenXpolynomials}
\alias{GreenXpolynomials}
\title{Green X-polynomials}
\usage{
GreenXpolynomials(rho)
}
\arguments{
\item{rho}{integer partition}
}
\value{
The Green X-polynomials, usually denoted by \eqn{X_\rho^\lambda}, 
  for all integer partitions \eqn{\lambda} of the same weight as the 
  integer partition \eqn{\rho}. 
  They are returned in a list. Each element of this list is itself a list, 
  with two elements. The first one, called \code{lambda}, represents the 
  partition \eqn{\lambda}. The second one, called \code{polynomial}, 
  represents the Green X-polynomial \eqn{X_\rho^\lambda}. This is a 
  univariate \code{qspray} polynomial whose variable is denoted by \code{t}.
  The names of the list encode the partitions \eqn{\lambda}.
}
\description{
Computes the Green X-polynomials for a given partition.
}
\examples{
GreenXpolynomials(c(2, 1)) 
}
