% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HallLittlewood.R
\name{HallLittlewood}
\alias{HallLittlewood}
\title{Hall-Littlewood polynomial}
\usage{
HallLittlewood(n, lambda, which = "P")
}
\arguments{
\item{n}{number of variables}

\item{lambda}{integer partition}

\item{which}{which Hall-Littlewood polynomial, \code{"P"} or \code{"Q"}}
}
\value{
The Hall-Littlewood polynomial in \code{n} variables of the
  integer partition \code{lambda}. This is a \code{symbolicQspray}
  polynomial with a unique parameter.
}
\description{
Hall-Littlewood polynomial of a given partition.
}
