% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLcombinations.R
\name{HLcombinationQ}
\alias{HLcombinationQ}
\title{Symmetric polynomial in terms of Hall-Littlewood 
  Q-polynomials}
\usage{
HLcombinationQ(Qspray, check = TRUE)
}
\arguments{
\item{Qspray}{a \code{qspray} polynomial or a 
\code{symbolicQspray} polynomial, which must be symmetric}

\item{check}{Boolean, whether to check the symmetry}
}
\value{
A list defining the linear combination. Each element of this list 
  is itself a list, with two elements: \code{coeff}, which is a 
  \code{ratioOfQsprays}, and the second element of this list is 
  \code{lambda}, an integer partition; then this list
  corresponds to the term \code{coeff * HallLittlewoodPol(n, lambda, "Q")} 
  of the linear combination, where \code{n} is the number of variables in 
  the symmetric polynomial \code{Qspray}. 
  The output list defining the linear combination is named 
  by some character strings representing the partitions \code{lambda}.
}
\description{
Returns the expression of a symmetric polynomial as a 
  linear combination of some Hall-Littlewood Q-polynomials.
}
