% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia.R
\name{Jack_julia}
\alias{Jack_julia}
\title{Evaluation with Julia}
\usage{
Jack_julia()
}
\value{
A function with the same arguments as .
}
\description{
Evaluate the hypergeometric function of a matrix argument with
  Julia. This is highly faster.
}
\note{
See \code{\link[JuliaConnectoR]{JuliaConnectoR-package}} for
  information about setting up Julia. If you want to directly use Julia,
  you can use \href{https://github.com/stla/HypergeoMat.jl}{my package}.
}
\examples{
library(jack)
\donttest{if(JuliaConnectoR::juliaSetupOk()){
  JuliaConnectoR::stopJulia()
}}
}
