% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polynomials.R
\name{JackPol}
\alias{JackPol}
\title{Jack polynomial}
\usage{
JackPol(m, lambda, alpha, basis = "canonical")
}
\arguments{
\item{m}{number of variables, a positive integer}

\item{lambda}{an integer partition, given as a vector of decreasing
integers}

\item{alpha}{parameter of the Jack polynomial,
a positive \code{bigq} rational number}

\item{basis}{the polynomial basis, either \code{"canonical"} or
\code{"MSF"} (monomial symmetric functions)}
}
\value{
A polynomial (\code{mvp} object; see \link[mvp]{mvp-package}).
}
\description{
Returns the Jack polynomial.
}
\examples{
JackPol(3, lambda = c(3,1), alpha = gmp::as.bigq(2,3))
JackPol(3, lambda = c(3,1), alpha= gmp::as.bigq(2,3), basis = "MSF")
}
