% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPP.R
\name{JackPolCPP}
\alias{JackPolCPP}
\title{Jack polynomial - C++ implementation}
\usage{
JackPolCPP(n, lambda, alpha)
}
\arguments{
\item{n}{number of variables, a positive integer}

\item{lambda}{an integer partition, given as a vector of decreasing
integers}

\item{alpha}{positive rational number, given as a string such as
\code{"2/3"} or as a \code{bigq} number}
}
\value{
A \code{qspray} multivariate polynomial.
}
\description{
Returns the Jack polynomial.
}
\examples{
JackPolCPP(3, lambda = c(3, 1), alpha = "2/5")
}
