% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jack.R
\name{JackQ}
\alias{JackQ}
\title{Evaluation of Jack polynomial with rational parameters}
\usage{
JackQ(x, lambda, alpha)
}
\arguments{
\item{x}{\link[gmp]{bigq} vector}

\item{lambda}{integer partition, a vector of decreasing integers}

\item{alpha}{positive \code{bigq} rational}
}
\value{
A \code{bigq} rational number.
}
\description{
Evaluates the Jack polynomials with rational parameters.
}
\examples{
x <- c(1, 2, 5/2)
lambda <- c(3, 1)
Jack(x, lambda, alpha = 2)
library(gmp)
x <- c(as.bigq(1), as.bigq(2), as.bigq(5,2))
JackQ(x, lambda, alpha = as.bigq(2))
}
