% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HallLittlewood.R
\name{KostaFoulkesPolynomial}
\alias{KostaFoulkesPolynomial}
\title{Kostka-Foulkes polynomial}
\usage{
KostaFoulkesPolynomial(lambda, mu)
}
\arguments{
\item{lambda, mu}{integer partitions}
}
\value{
The Kostka-Foulkes polynomial associated to \code{lambda} and
  \code{mu}. This is a univariate \code{qspray} polynomial.
}
\description{
Kostka-Foulkes polynomial for two given partitions.
}
