% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kostka.R
\name{KostkaJackNumbers}
\alias{KostkaJackNumbers}
\title{Kostka-Jack numbers}
\usage{
KostkaJackNumbers(n, alpha = NULL)
}
\arguments{
\item{n}{positive integer, the weight of the partitions}

\item{alpha}{the Jack parameter, a \code{bigq} number or an object coercible
to a \code{bigq} number; setting \code{alpha=NULL} is equivalent to set
\code{alpha=1}}
}
\value{
A matrix of character strings representing integers or fractions.
}
\description{
The Kostka-Jack numbers for partitions of a given weight.
}
\details{
The Kostka-Jack numbers generalize the Kostka numbers: these ones
  are obtained when \code{alpha=1}.
}
\examples{
KostkaJackNumbers(4)
}
