% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSF.R
\name{MSFpoly}
\alias{MSFpoly}
\title{Monomial symmetric function}
\usage{
MSFpoly(m, lambda)
}
\arguments{
\item{m}{integer, the number of variables}

\item{lambda}{an integer partition, given as a vector of decreasing
integers}
}
\value{
A polynomial (\code{mvp} object; see \link[mvp]{mvp-package}).
}
\description{
Returns a monomial symmetric function as a polynomial.
}
\examples{
MSFpoly(3, c(3,1))
}
