% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msPolynomial.R
\name{msPolynomial}
\alias{msPolynomial}
\title{Monomial symmetric polynomial}
\usage{
msPolynomial(n, lambda)
}
\arguments{
\item{n}{integer, the number of variables}

\item{lambda}{an integer partition, given as a vector of decreasing
nonnegative integers}
}
\value{
A \code{qspray} object.
}
\description{
Returns a monomial symmetric polynomial.
}
\examples{
library(jack)
msPolynomial(3, c(3, 1))
}
