% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia.R
\name{print.exactmvp}
\alias{print.exactmvp}
\title{Print an \code{exactmvp} object}
\usage{
\method{print}{exactmvp}(x, ...)
}
\arguments{
\item{x}{object of class \code{exactmvp}; the functions returned by
\code{\link{Jack_julia}} can return such objects}

\item{...}{arguments passed to \code{\link[mvp]{print.mvp}}}
}
\value{
Nothing.
}
\description{
Print an \code{exactmvp} object.
}
