% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicSymmetricPolynomials.R
\name{psPolynomial}
\alias{psPolynomial}
\title{Power sum polynomial}
\usage{
psPolynomial(n, lambda)
}
\arguments{
\item{n}{integer, the number of variables}

\item{lambda}{an integer partition, given as a vector of decreasing
positive integers}
}
\value{
A \code{qspray} object.
}
\description{
Returns a power sum polynomial.
}
\examples{
library(jack)
psPolynomial(3, c(3, 1))
}
