% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{roll}
\alias{roll}
\title{Dice}
\usage{
roll(dietype = 1:6, probvalues = rep(1/length(dietype), length(dietype)))
}
\arguments{
\item{dietype}{The type of die}

\item{probvalues}{The probability of getting a particular number.}
}
\value{
the sum of two dice
}
\description{
Dice
}
\examples{
set.seed(123)
roll()
roll(1:20)
}

