% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{RP}
\alias{RP}
\title{Power Function}
\usage{
RP(x, power = 1)
}
\arguments{
\item{x}{Must be a numeric object}

\item{power}{A numeric object called power \code{power}}
}
\value{
Raises the values in the numeric object \code{x} to the value in \code{power}
}
\description{
Power Function
}
\examples{
RP(c(1, 3, 5), 3)
RP(x = c(1, 3, 5), power = 3)
}

