% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schur.R
\name{Schur}
\alias{Schur}
\title{Schur polynomial}
\usage{
Schur(x, lambda)
}
\arguments{
\item{x}{numeric vector}

\item{lambda}{integer partition}
}
\value{
A number.
}
\description{
Evaluates the Schur polynomials.
}
\examples{
x <- c(2,3,4)
Schur(x, c(2,1,1))
prod(x) * sum(x)
}
