% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPP.R
\name{SchurPolCPP}
\alias{SchurPolCPP}
\title{Schur polynomial - C++ implementation}
\usage{
SchurPolCPP(n, lambda)
}
\arguments{
\item{n}{number of variables, a positive integer}

\item{lambda}{an integer partition, given as a vector of decreasing
integers}
}
\value{
A \code{qspray} multivariate polynomial.
}
\description{
Returns the Schur polynomial.
}
\examples{
SchurPolCPP(3, lambda = c(3, 1))
}
