% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{shuffle}
\alias{shuffle}
\title{Everyday I'm shuffling}
\usage{
shuffle(deck)
}
\arguments{
\item{cards}{}
}
\value{
shuffles the deck to prevent same outcome each time
}
\description{
Everyday I'm shuffling
}
\examples{
site <- "https://gist.githubusercontent.com/garrettgman/9629323/raw/ee5dfc039fd581cb467cc69c226ea2524913c3d8/deck.csv"
deck <- readr::read_csv(site)
write.csv(deck, file = "cards.csv", row.names = FALSE)
deck[1, c("face", "suit", "value")]
deal <- function(deck) {
deck[1,1]
}
}

