% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkewHallLittlewood.R
\name{SkewHallLittlewood}
\alias{SkewHallLittlewood}
\title{Skew Hall-Littlewood polynomial}
\usage{
SkewHallLittlewood(n, lambda, mu, which = "P")
}
\arguments{
\item{n}{number of variables, a positive integer}

\item{lambda, mu}{integer partitions defining the skew partition:
\code{lambda} is the outer partition and \code{mu} is the inner partition
(so \code{mu} must be a subpartition of \code{lambda})}

\item{which}{which Hall-Littlewood polynomial, \code{"P"} or \code{"Q"}}
}
\value{
A \code{symbolicQspray} multivariate polynomial, the skew
  Hall-Littlewood polynomial associated to the skew partition defined by
  \code{lambda} and \code{mu}. It has only one parameter and its
  coefficients are polynomial in this parameter.
}
\description{
Returns the skew
}
\examples{
n <- 3; lambda <- c(3, 2, 1); mu <- c(1, 1)
skewHLpoly <- SkewHallLittlewood(n, lambda, mu)
skewSchurPoly <- SkewSchurPol(n, lambda, mu)
substituteParameters(skewHLpoly, 0) == skewSchurPoly # should be TRUE
}
