% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkewMacdonaldPolynomial.R
\name{SkewMacdonaldPol}
\alias{SkewMacdonaldPol}
\title{Skew Macdonald polynomial}
\usage{
SkewMacdonaldPol(n, lambda, mu, which)
}
\arguments{
\item{n}{number of variables, a positive integer}

\item{lambda, mu}{integer partitions defining the skew partition:
\code{lambda} is the outer partition and \code{mu} is the inner partition
(so \code{mu} must be a subpartition of \code{lambda})}

\item{which}{which skew Macdonald polynomial, \code{"P"}, \code{"Q"}
or \code{"J"}}
}
\value{
A \code{symbolicQspray} multivariate polynomial, the skew
  Macdonald polynomial associated to the skew partition defined by
  \code{lambda} and \code{mu}. It has two parameters.
}
\description{
Returns the skew Macdonald polynomial associated to
  the given skew partition.
}
