% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSchurPolynomial.R
\name{tSchurPol}
\alias{tSchurPol}
\title{t-Schur polynomial}
\usage{
tSchurPol(n, lambda)
}
\arguments{
\item{n}{number of variables, a positive integer}

\item{lambda}{integer partition}
}
\value{
A \code{symbolicQspray} multivariate polynomial, the
  t-Schur polynomial associated to \code{lambda}.
  It has one parameter and its coefficients
  are polynomials in this parameter.
}
\description{
Returns the t-Schur polynomial associated to
  the given partition.
}
