% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonal.R
\name{Zonal}
\alias{Zonal}
\title{Zonal polynomial}
\usage{
Zonal(x, lambda)
}
\arguments{
\item{x}{numeric vector}

\item{lambda}{integer partition}
}
\value{
A number.
}
\description{
Evaluates the zonal polynomials.
}
\examples{
x <- c(3,1)
Zonal(x, c(1,1)) + Zonal(x, 2) # sum(x)^2
Zonal(x, 3) + Zonal(x, c(2,1)) + Zonal(x, c(1,1,1)) # sum(x)^3
}
