% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPP.R
\name{ZonalPolCPP}
\alias{ZonalPolCPP}
\title{Zonal polynomial - C++ implementation}
\usage{
ZonalPolCPP(m, lambda)
}
\arguments{
\item{m}{number of variables, a positive integer}

\item{lambda}{an integer partition, given as a vector of decreasing
integers}
}
\value{
A \code{qspray} multivariate polynomial.
}
\description{
Returns the Zonal polynomial.
}
\examples{
ZonalPolCPP(3, lambda = c(3, 1))
}
