% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPP.R
\name{ZonalQPolCPP}
\alias{ZonalQPolCPP}
\title{Quaternionic zonal polynomial - C++ implementation}
\usage{
ZonalQPolCPP(m, lambda)
}
\arguments{
\item{m}{number of variables, a positive integer}

\item{lambda}{an integer partition, given as a vector of decreasing
integers}
}
\value{
A \code{qspray} multivariate polynomial.
}
\description{
Returns the quaternionic zonal polynomial.
}
\examples{
ZonalQPolCPP(3, lambda = c(3, 1))
}
