% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jack.R
\name{Jack}
\alias{Jack}
\title{Jack polynomial}
\usage{
Jack(x, lambda, alpha)
}
\arguments{
\item{x}{numeric vector}

\item{lambda}{integer partition, a vector of decreasing integers}

\item{alpha}{positive number}
}
\value{
A number.
}
\description{
Evaluates the Jack polynomials.
}
