% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kostka.R
\name{KostkaNumbers}
\alias{KostkaNumbers}
\title{Kostka numbers}
\usage{
KostkaNumbers(n)
}
\arguments{
\item{n}{positive integer, the weight of the partitions}
}
\value{
A matrix of integers.
}
\description{
The Kostka numbers for partitions of a given weight.
}
\examples{
KostkaNumbers(4)
}
