% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schur.R
\name{Schur}
\alias{Schur}
\title{Evaluation of Schur polynomials}
\usage{
Schur(x, lambda, algorithm = "DK")
}
\arguments{
\item{x}{numeric vector or \link[gmp]{bigq} vector}

\item{lambda}{an integer partition, given as a vector of decreasing
integers}

\item{algorithm}{the algorithm used, either \code{"DK"} (Demmel-Koev)
or \code{"naive"}}
}
\value{
A number or a \code{bigq} rational number.
}
\description{
Evaluates the Schur polynomials.
}
\examples{
x <- c(2,3,4)
Schur(x, c(2,1,1))
prod(x) * sum(x)
}
\seealso{
\code{\link{SchurPol}}
}
