% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polynomials.R
\name{ZonalPol}
\alias{ZonalPol}
\title{Zonal polynomial}
\usage{
ZonalPol(m, lambda, basis = "canonical")
}
\arguments{
\item{m}{number of variables, a positive integer}

\item{lambda}{an integer partition, given as a vector of decreasing
integers}

\item{basis}{the polynomial basis, either \code{"canonical"} or
\code{"MSF"} (monomial symmetric functions)}
}
\value{
A polynomial (\code{mvp} object; see \link[mvp]{mvp-package}).
}
\description{
Returns the Zonal polynomial.
}
\examples{
ZonalPol(3, lambda = c(3,1))
ZonalPol(3, lambda = c(3,1), basis = "MSF")
}
