% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polynomials.R
\name{ZonalQPol}
\alias{ZonalQPol}
\title{Quaternionic zonal polynomial}
\usage{
ZonalQPol(m, lambda, basis = "canonical", exact = TRUE)
}
\arguments{
\item{m}{number of variables, a positive integer}

\item{lambda}{an integer partition, given as a vector of decreasing
integers}

\item{basis}{the polynomial basis, either \code{"canonical"} or
\code{"MSF"} (monomial symmetric functions)}

\item{exact}{logical, whether to use exact arithmetic to calculate the
coefficients}
}
\value{
A polynomial (\code{mvp} object; see \link[mvp]{mvp-package}).
}
\description{
Returns the quaternionic zonal polynomial.
}
\examples{
ZonalQPol(3, lambda = c(3,1))
ZonalQPol(3, lambda = c(3,1), basis = "MSF")
}
